requirements.txt
addict==2.4.0
airsim==1.7.0
attr==0.3.2
decord==0.6.0
deepspeed==0.17.2
einops==0.8.1
fastdtw==0.3.4
flash_attn==2.8.3
huggingface_hub==0.33.4
hydra-core==1.3.2
ipdb==0.13.13
matplotlib==3.10.7
MatterSim==1.2.0
msgpack_rpc_python==0.4.1
networkx==3.2.1
numba==0.60.0
numpy==2.3.5
opencv_contrib_python==4.12.0.88
opencv_python==4.11.0.86
packaging==25.0
peft==0.10.0
pickle5==0.0.12
Pillow==12.0.0
pycocotools==2.0.10
PyYAML==6.0.2
Requests==2.32.5
safetensors==0.7.0
scipy==1.16.3
setuptools==78.1.1
supervision==0.27.0
tensordict==0.4.0
termcolor==3.2.0
tf==1.1.0
timm==0.9.12
tqdm==4.67.1
transformers==4.53.0.dev0
transformers.egg==info
wandb==0.23.0
yacs==0.1.8
yapf==0.43.0
